requirements.txt
python-chess==0.28.2
numpy==1.17.1
tensorflow==1.14.0
tflearn==0.3.2
setuptools